+++ src/CallbackOption.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Jeffreyvr\WPMetaBox;
+
+class CallbackOption
+{
+    public \Closure $callback;
+
+    public function __construct($callback)
+    {
+        $this->callback = $callback;
+    }
+
+    public function render()
+    {
+        $callback = $this->callback;
+
+        return $callback();
+    }
+}